Refactor hk to own dir
