+++ cg/cli/workflow/mip_dna/base.py
@@ -5,7 +5,7 @@ import logging
 import click
 
 from cg.apps.environ import environ_email
-from cg.apps.hk import HousekeeperAPI
+from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
 from cg.apps.scout.scoutapi import ScoutAPI
 from cg.apps.tb import TrailblazerAPI
