+++ cg/meta/workflow/mip.py
@@ -7,7 +7,7 @@ from typing import Any, List
 
 from ruamel.yaml import safe_load
 
-from cg.apps.hk import HousekeeperAPI
+from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
 from cg.apps.mip import parse_trending
 from cg.apps.mip.base import MipAPI
@@ -15,14 +15,8 @@ from cg.apps.mip.confighandler import ConfigHandler
 from cg.apps.scout.scoutapi import ScoutAPI
 from cg.apps.tb import TrailblazerAPI
 from cg.apps.tb.models import TrailblazerAnalysis
-from cg.constants import (
-    COLLABORATORS,
-    COMBOS,
-    DEFAULT_CAPTURE_KIT,
-    CASE_ACTIONS,
-    MASTER_LIST,
-    Pipeline,
-)
+from cg.constants import (CASE_ACTIONS, COLLABORATORS, COMBOS,
+                          DEFAULT_CAPTURE_KIT, MASTER_LIST, Pipeline)
 from cg.exc import CgDataError, LimsDataError
 from cg.store import Store, models
