+++ tests/cli/clean/conftest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 from cg.apps.balsamic.api import BalsamicAPI
-from cg.apps.hk import HousekeeperAPI
+from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.constants import Pipeline
 from cg.meta.workflow.balsamic import BalsamicAnalysisAPI
 from cg.store import Store
