+++ tests/cli/upload/conftest.py
@@ -1,20 +1,18 @@
 """Fixtures for cli balsamic tests"""
 import json
 import logging
-from pathlib import Path
 from datetime import datetime
+from pathlib import Path
 
 import pytest
 
-from cg.apps.hk import HousekeeperAPI
-from cg.apps.scout.scoutapi import ScoutAPI
 from cg.apps.gt import GenotypeAPI
+from cg.apps.housekeeper.hk import HousekeeperAPI
+from cg.apps.scout.scout_load_config import ScoutLoadConfig
+from cg.apps.scout.scoutapi import ScoutAPI
 from cg.meta.upload.scoutapi import UploadScoutAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
-from cg.store import Store
-from cg.store import models
-from cg.apps.scout.scout_load_config import ScoutLoadConfig
-
+from cg.store import Store, models
 from tests.mocks.madeline import MockMadelineAPI
 
 LOG = logging.getLogger(__name__)
