+++ tests/meta/conftest.py
@@ -5,13 +5,12 @@ from datetime import datetime
 import pytest
 from _pytest import tmpdir
 
-from cg.store import Store
 from cg.apps.balsamic.fastq import FastqHandler as BalsamicFastqHandler
+from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.microsalt.fastq import FastqHandler as MicrosaltFastqHandler
-from cg.apps.hk import HousekeeperAPI
 from cg.apps.scout.scoutapi import ScoutAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
-
+from cg.store import Store
 from tests.store_helpers import StoreHelpers
